Add license header in new files
