--- Bemol/Onboarding/TipHandler.swift
@@ -1,6 +1,0 @@
-import Foundation
-
-@MainActor
-protocol TipHandler {
-  func handle(_ tip: Tip)
-}
