+++ Bemol/Views/Tip/TipPresenter.swift
@@ -1,5 +1,22 @@
-import UIKit
+///
+/// TipPresenter.swift
+/// Bemol
+///
+/// Copyright 2025 Faiçal Tchirou
+///
+/// Bemol is free software: you can redistribute it and/or modify it under the terms of
+/// the GNU General Public License as published by the Free Software Foundation, either version 3
+/// of the License, or (at your option) any later version.
+///
+/// Bemol is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY;
+/// without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
+/// See the GNU General Public License for more details.
+///
+/// You should have received a copy of the GNU General Public License along with Foobar.
+/// If not, see <https://www.gnu.org/licenses/>.
+///
 
+import UIKit
 
 @MainActor
 final class TipPresenter {
@@ -107,6 +124,7 @@ final class TipPresenter {
         if tipView.superview != nil {
           tipView.removeFromSuperview()
         }
+  
         completion()
       }
     )
